rsa-keyGen/test/NIST.CVP.Generation.RSA-KeyGen.Tests/TestCaseGeneratorFactoryTests.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using NIST.CVP.Crypto.RSA;
using NIST.CVP.Tests.Core.TestCategoryAttributes;
using NUnit.Framework;

namespace NIST.CVP.Generation.RSA_KeyGen.Tests
{
    [TestFixture, UnitTest]
    public class TestCaseGeneratorFactoryTests
    {
        [Test]
        [TestCase(KeyGenModes.B32, "aft", typeof(TestCaseGeneratorAFT_B32))]
        public void ShouldReturnProperGenerator(KeyGenModes mode, string testType, Type expectedType)
        {
            var testGroup = new TestGroup
            {
                Mode = mode,
                TestType = testType
            };

            var subject = new TestCaseGeneratorFactory(null);
            var generator = subject.GetCaseGenerator(testGroup);
            Assume.That(generator != null);
            Assert.IsInstanceOf(expectedType, generator);
        }
    }
}
